Developer

+++ src/CarPlates.API/Models/ItemBarCodeModels.cs
@@ -34,7 +34,7 @@ public class ItemBarCodeView
     public string? ItemDiscount3 { get; set; }
     public byte? Status { get; set; }
     public double? Pkg2Qty { get; set; }
-    public double? Pkg3Qty { get; set; }
+    public int? Pkg3Qty { get; set; }
     public double? Pkg1Price1 { get; set; }
     public double? Pkg2Price1 { get; set; }
     public double? Pkg3Price1 { get; set; }
